Apply improved types & architecture to Stellar
