+++ libs/ledger-live-common/src/families/stellar/bridge/js.ts
@@ -1,32 +1,22 @@
-import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
-import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
+import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
+import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
+import estimateMaxSpendable from "../estimateMaxSpendable";
+import getTransactionStatus from "../getTransactionStatus";
+import { scanAccounts, sync } from "../synchronization";
+import prepareTransaction from "../prepareTransaction";
+import createTransaction from "../createTransaction";
+import signOperation from "../signOperation";
 import type { Transaction } from "../types";
-import { scanAccounts, sync } from "../js-synchronization";
-import getTransactionStatus from "../js-getTransactionStatus";
-import estimateMaxSpendable from "../js-estimateMaxSpendable";
-import createTransaction from "../js-createTransaction";
-import prepareTransaction from "../js-prepareTransaction";
-import signOperation from "../js-signOperation";
-import broadcast from "../js-broadcast";
-import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
-
-const preload = async (): Promise<any> => {};
-
-const hydrate = (): void => {};
-
-const receive = makeAccountBridgeReceive();
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const getPreloadStrategy = (currency: CryptoCurrency): any => ({});
+import broadcast from "../broadcast";
 
 const currencyBridge: CurrencyBridge = {
-  getPreloadStrategy,
-  preload,
-  hydrate,
+  preload: async () => {},
+  hydrate: () => {},
   scanAccounts,
 };
 
+const receive = makeAccountBridgeReceive();
 const accountBridge: AccountBridge<Transaction> = {
   createTransaction,
   updateTransaction: defaultUpdateTransaction,
