+++ libs/ledger-live-common/src/families/stellar/buildOptimisticOperation.ts
@@ -1,0 +1,64 @@
+import BigNumber from "bignumber.js";
+import { Account } from "@ledgerhq/types-live";
+import { StellarOperation, Transaction } from "./types";
+import { encodeOperationId } from "../../operation";
+import { getAmountValue } from "./logic";
+import { fetchSequence } from "./api";
+
+export const buildOptimisticOperation = async (
+  account: Account,
+  transaction: Transaction,
+): Promise<StellarOperation> => {
+  const transactionSequenceNumber = await fetchSequence(account);
+  const fees = transaction.fees ?? new BigNumber(0);
+  const type = transaction.mode === "changeTrust" ? "OPT_IN" : "OUT";
+
+  const operation: StellarOperation = {
+    id: encodeOperationId(account.id, "", type),
+    hash: "",
+    type,
+    value: transaction.subAccountId ? fees : getAmountValue(account, transaction, fees),
+    fee: fees,
+    blockHash: null,
+    blockHeight: null,
+    senders: [account.freshAddress],
+    recipients: [transaction.recipient],
+    accountId: account.id,
+    date: new Date(),
+    // FIXME: Javascript number may be not precise enough
+    transactionSequenceNumber: transactionSequenceNumber?.plus(1).toNumber(),
+    extra: {
+      ledgerOpType: type,
+    },
+  };
+
+  const { subAccountId } = transaction;
+  const { subAccounts } = account;
+
+  const tokenAccount = !subAccountId
+    ? null
+    : subAccounts && subAccounts.find(ta => ta.id === subAccountId);
+
+  if (tokenAccount && subAccountId) {
+    operation.subOperations = [
+      {
+        id: `${subAccountId}--OUT`,
+        hash: "",
+        type: "OUT",
+        value: transaction.useAllAmount ? tokenAccount.balance : transaction.amount,
+        fee: new BigNumber(0),
+        blockHash: null,
+        blockHeight: null,
+        senders: [account.freshAddress],
+        recipients: [transaction.recipient],
+        accountId: subAccountId,
+        date: new Date(),
+        extra: {
+          ledgerOpType: type,
+        },
+      },
+    ];
+  }
+
+  return operation;
+};
