+++ libs/ledger-live-common/src/families/stellar/createTransaction.ts
@@ -1,12 +1,13 @@
 import { BigNumber } from "bignumber.js";
 import type { Transaction } from "./types";
+import { AccountBridge } from "@ledgerhq/types-live";
 
 /**
  * Create an empty transaction
  *
  * @returns {Transaction}
  */
-const createTransaction = (): Transaction => ({
+const createTransaction: AccountBridge<Transaction>["createTransaction"] = () => ({
   family: "stellar",
   amount: new BigNumber(0),
   baseReserve: null,
