+++ libs/ledger-live-common/src/families/stellar/prepareTransaction.ts
@@ -1,0 +1,35 @@
+import invariant from "invariant";
+import type { AccountBridge } from "@ledgerhq/types-live";
+import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
+import { fetchAccountNetworkInfo } from "./api";
+import { getAssetCodeIssuer } from "./logic";
+import type { Transaction } from "./types";
+
+const prepareTransaction: AccountBridge<Transaction>["prepareTransaction"] = async (
+  account,
+  transaction,
+) => {
+  const networkInfo = transaction.networkInfo || (await fetchAccountNetworkInfo(account));
+  invariant(networkInfo.family === "stellar", "stellar networkInfo expected");
+  const fees = transaction.fees || networkInfo.fees;
+  const baseReserve = transaction.baseReserve || networkInfo.baseReserve;
+  const [assetCode, assetIssuer] = getAssetCodeIssuer(transaction);
+
+  if (
+    transaction.networkInfo !== networkInfo ||
+    transaction.fees !== fees ||
+    transaction.baseReserve !== baseReserve
+  ) {
+    return defaultUpdateTransaction(transaction, {
+      networkInfo,
+      fees,
+      baseReserve,
+      assetCode,
+      assetIssuer,
+    });
+  }
+
+  return defaultUpdateTransaction(transaction, { assetCode, assetIssuer });
+};
+
+export default prepareTransaction;
