+++ libs/ledger-live-common/src/families/stellar/synchronization.ts
